# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A book management service API in Go using Gin framework.",
        "title": "Gin Book Service",
        "termsOfService": "https://tos.santoshk.dev",
        "contact": {
            "name": "Santosh Kumar",
            "url": "https://twitter.com/sntshk",
            "email": "sntshkmr60@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/books": {
            "get": {
                "description": "Responds with the list of all books as JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Get books array",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Takes a book JSON and store in DB. Return saved JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Store a new book",
                "parameters": [
                    {
                        "description": "Book JSON",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                }
            }
        },
        "/books/{isbn}": {
            "get": {
                "description": "Returns the book whose ISBN value matches the isbn.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Get single book by isbn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search book by isbn",
                        "name": "isbn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Book": {
            "type": "object",
            "properties": {
                "bar": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 4,
                    "example": "random string"
                },
                "baz": {
                    "type": "integer",
                    "default": 15,
                    "maximum": 20,
                    "minimum": 10
                },
                "qux": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "foo",
                            "bar",
                            "baz"
                        ]
                    }
                }
            }
        }
    }
}